requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
easypikaio==0.0.3
paho-mqtt==1.6.1
pika==1.2.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
pyvim==3.0.3
six==1.16.0
sqlparse==0.4.2
wcwidth==0.2.5
